pyproject.toml
[project]
name = "fastapi-pcr"
version = "0.1.0"
description = ""
authors = [
    {name = "mayconcorreatinti",email = "maycon326754@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi (>=0.119.0,<0.120.0)",
    "aiomysql (>=0.2.0,<0.3.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "mysql-connector-python (>=9.4.0,<10.0.0)",
    "pwdlib[argon2] (>=0.3.0,<0.4.0)",
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
